ApplicationPlayer: relocate to application directory
