+++ xbmc/application/ApplicationPlayer.cpp
