+++ xbmc/application/ApplicationPlayer.h
