+++ xbmc/cores/RetroPlayer/guiwindows/GameWindowFullScreen.cpp
@@ -9,10 +9,10 @@
 #include "GameWindowFullScreen.h"
 
 #include "Application.h" //! @todo Remove me
-#include "ApplicationPlayer.h" //! @todo Remove me
 #include "GUIInfoManager.h" //! @todo Remove me
 #include "GameWindowFullScreenText.h"
 #include "ServiceBroker.h"
+#include "application/ApplicationPlayer.h" //! @todo Remove me
 #include "cores/RetroPlayer/guibridge/GUIGameRenderManager.h"
 #include "cores/RetroPlayer/guibridge/GUIRenderHandle.h"
 #include "games/GameServices.h"
